Renamed the remote logging components
