+++ components/logging/org.wso2.carbon.logging.remote.config.ui/src/main/resources/META-INF/component.xml
@@ -1,0 +1,35 @@
+<!--
+ ~ Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ ~
+ ~ WSO2 Inc. licenses this file to you under the Apache License,
+ ~ Version 2.0 (the "License"); you may not use this file except
+ ~ in compliance with the License.
+ ~ You may obtain a copy of the License at
+ ~
+ ~ http://www.apache.org/licenses/LICENSE-2.0
+ ~
+ ~ Unless required by applicable law or agreed to in writing,
+ ~ software distributed under the License is distributed on an
+ ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ ~ KIND, either express or implied. See the License for the
+ ~ specific language governing permissions and limitations
+ ~ under the License.
+-->
+<component xmlns="http://products.wso2.org/carbon">
+        <!-- sample menu configuration -->
+    <menus>
+        <menu>
+            <id>logging_config_menu</id>
+            <i18n-key>log-config.logs</i18n-key>
+            <i18n-bundle>org.wso2.carbon.logging.remote.config.ui.i18n.Resources</i18n-bundle>
+            <parent-menu>configure_menu</parent-menu>
+            <link>../log-config/log-config.jsp</link>
+            <region>region1</region>
+            <order>12</order>
+            <style-class>monitor</style-class>
+            <icon>../log-config/images/log.gif</icon>
+            <require-permission>/permission/protected/configure/logging</require-permission>
+            <require-super-tenant>true</require-super-tenant>
+        </menu>
+    </menus>
+</component>
