+++ components/logging/org.wso2.carbon.logging.remote.config.ui/src/main/resources/org/wso2/carbon/logging/remote/config/ui/i18n/Resources.properties
