+++ components/logging/org.wso2.carbon.logging.remote.config.ui/src/main/resources/web/log-config/images/log.gif
