+++ components/logging/org.wso2.carbon.logging.remote.config.ui/src/main/resources/web/log-config/js/loggingconfig.js
