+++ components/logging/org.wso2.carbon.logging.service/src/main/resources/META-INF/services.xml
@@ -28,7 +28,7 @@
         <parameter name="AuthorizationAction" locked="true">/permission/protected/configure/logging</parameter>
         <parameter name="enableMTOM" locked="false">true</parameter>
     </service>
-    <service name="LoggingConfig" scope="transportsession">
+    <service name="RemoteLoggingConfig" scope="transportsession">
         <schema schemaNamespace="http://org.apache.axis2/xsd" elementFormDefaultQualified="true"/>
         <transports>
             <transport>https</transport>
@@ -36,7 +36,7 @@
         <description>
             This service is used for configuring the remote server logging configuration
         </description>
-        <parameter name="ServiceClass">org.wso2.carbon.logging.service.LoggingConfig</parameter>
+        <parameter name="ServiceClass">org.wso2.carbon.logging.service.RemoteLoggingConfig</parameter>
         <parameter name="AuthorizationAction" locked="true">/permission/protected/configure/logging</parameter>
         <parameter name="enableMTOM" locked="false">true</parameter>
     </service>
