+++ features/logging-mgt/pom.xml
@@ -35,8 +35,8 @@
         <module>org.wso2.carbon.logging.viewer.feature</module>
         <module>org.wso2.carbon.logging.mgt.ui.feature</module>
         <module>org.wso2.carbon.logging.mgt.feature</module>
-        <module>org.wso2.carbon.logging.config.mgt.ui.feature</module>
-        <module>org.wso2.carbon.logging.config.mgt.feature</module>
+        <module>org.wso2.carbon.logging.remote.config.mgt.ui.feature</module>
+        <module>org.wso2.carbon.logging.remote.config.mgt.feature</module>
     </modules>
