+++ service-stubs/logging/pom.xml
@@ -34,7 +34,7 @@
     <modules>
         <module>org.wso2.carbon.logging.view.stub</module>
         <module>org.wso2.carbon.logging.admin.stub</module>
-        <module>org.wso2.carbon.logging.config.stub</module>
+        <module>org.wso2.carbon.logging.remote.config.stub</module>
     </modules>
 
     <build>
